pyproject.toml
[project]
name = "scope-ltx-2"
version = "0.3.0"
description = "LTX 2.3 audio-video generation plugin for Daydream Scope"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torchaudio==2.9.1",
    "safetensors",
    "einops",
    "transformers",
]

[project.entry-points.scope]
scope-ltx-2 = "scope_ltx_2"

[dependency-groups]
dev = [
    "daydream-scope",
]

[tool.uv]
override-dependencies = [
]

[tool.uv.sources]
daydream-scope = { git = "https://github.com/daydreamlive/scope" }
torchaudio = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
